Feat: 5주차 master branch merge

+++ BackEnd/globalcc/src/main/java/com/ssafy/globalcc/domain/user/service/UserServiceImpl.java
@@ -1,13 +1,17 @@
 package com.ssafy.globalcc.domain.user.service;
 
+import com.ssafy.globalcc.domain.team.entity.Team;
+import com.ssafy.globalcc.domain.team.repository.TeamRepository;
 import com.ssafy.globalcc.domain.user.dto.UserDto;
 import com.ssafy.globalcc.domain.user.entity.User;
 import com.ssafy.globalcc.domain.user.entity.UserDetail;
+import com.ssafy.globalcc.domain.user.exception.NoSuchUserException;
 import com.ssafy.globalcc.domain.user.repository.UserDetailRepository;
 import com.ssafy.globalcc.domain.user.repository.UserRepository;
 import com.ssafy.globalcc.domain.user.repository.UserTeamRepository;
 import com.ssafy.globalcc.domain.user.result.*;
 import com.ssafy.globalcc.utils.JwtUtil;
+import io.netty.util.internal.SocketUtils;
 import io.netty.util.internal.StringUtil;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
@@ -35,6 +39,7 @@ public class UserServiceImpl implements UserService {
     private final RedisOperations<String, String> redisOperations;
     @Value("${file.default}")
     private String fileDefault;
+    private final TeamRepository teamRepository;
 
     @Override
     public boolean checkDuplicateUserById(String id) {
@@ -62,6 +67,7 @@ public class UserServiceImpl implements UserService {
     }
 
     @Override
+    @Transactional
     public boolean logoutUser(int userId) {
         // Redis에서 refresh token 삭제
         var result = redisOperations.delete(Integer.toString(userId));
@@ -132,9 +138,27 @@ public class UserServiceImpl implements UserService {
     }
 
     @Override
+    @Transactional
     public void deleteUser(int userId) {
          // TODO: 정책을 참고해서 유저 삭제 진행
-        userRepository.deleteById(userId);
+        // 해당 유저가 팀장으로 있는 팀 조회
+        List<Team> teams = teamRepository.findAllByOwnerUserId(userId);
+        for (Team team : teams) {
+            // 같은 팀 팀원에게 팀장 권한 위임
+            List<UserTeam> userTeams = userTeamRepository.findAllByTeamTeamId(team.getTeamId());
+            if (!userTeams.isEmpty()) {
+                // 첫 번째 팀원을 새로운 팀장으로 지정
+                User newOwner = userTeams.get(0).getUser();
+                // 팀장 변경
+                team.setOwner(newOwner);
+                // 팀 변경 저장
+                teamRepository.save(team);
+            }
+        }
+        // 팀원인 경우
+        if (teams.isEmpty()) {
+            userRepository.deleteById(userId);
+        }
     }
 
     @Override
@@ -149,5 +173,11 @@ public class UserServiceImpl implements UserService {
         // 토큰에서 userID 추출
 
         return jwtUtil.createAccessToken(id);
+    }
+
+    @Override
+    public User findUserById(int userId) {
+        return userRepository.findUserByUserId(userId)
+                .orElseThrow(() -> new NoSuchUserException("User with id " + userId + " not found"));
     }
 }
